Reduce size of location expression arguments

+++ pkg/dynamicinstrumentation/ditypes/analysis.go
@@ -243,9 +243,9 @@ func DirectReadLocationExpression(p *Parameter) LocationExpression {
 		return LocationExpression{Opcode: OpInvalid}
 	}
 	if p.Location.InReg {
-		return ReadRegisterLocationExpression(uint(p.Location.Register), uint(p.TotalSize))
+		return ReadRegisterLocationExpression(uint16(p.Location.Register), uint16(p.TotalSize))
 	}
-	return ReadStackLocationExpression(uint(p.Location.StackOffset), uint(p.TotalSize))
+	return ReadStackLocationExpression(uint16(p.Location.StackOffset), uint16(p.TotalSize))
 }
 
 // DirectReadToOutputLocationExpression creates an expression which
@@ -256,9 +256,9 @@ func DirectReadToOutputLocationExpression(p *Parameter) LocationExpression {
 		return LocationExpression{Opcode: OpInvalid}
 	}
 	if p.Location.InReg {
-		return ReadRegisterToOutputLocationExpression(uint(p.Location.Register), uint(p.TotalSize))
+		return ReadRegisterToOutputLocationExpression(uint16(p.Location.Register), uint16(p.TotalSize))
 	}
-	return ReadStackToOutputLocationExpression(uint(p.Location.StackOffset), uint(p.TotalSize))
+	return ReadStackToOutputLocationExpression(uint16(p.Location.StackOffset), uint16(p.TotalSize))
 }
 
 // ReadRegisterLocationExpression creates an expression which
@@ -266,7 +266,7 @@ func DirectReadToOutputLocationExpression(p *Parameter) LocationExpression {
 // the top of the BPF parameter stack.
 // Arg1 = register
 // Arg2 = size of element
-func ReadRegisterLocationExpression(register, size uint) LocationExpression {
+func ReadRegisterLocationExpression(register, size uint16) LocationExpression {
 	return LocationExpression{Opcode: OpReadUserRegister, Arg1: register, Arg2: size}
 }
 
@@ -275,7 +275,7 @@ func ReadRegisterLocationExpression(register, size uint) LocationExpression {
 // into a u64 which is then pushed to the top of the BPF parameter stack.
 // Arg1 = stack offset
 // Arg2 = size of element
-func ReadStackLocationExpression(offset, size uint) LocationExpression {
+func ReadStackLocationExpression(offset, size uint16) LocationExpression {
 	return LocationExpression{Opcode: OpReadUserStack, Arg1: offset, Arg2: size}
 }
 
@@ -284,7 +284,7 @@ func ReadStackLocationExpression(offset, size uint) LocationExpression {
 // the output buffer.
 // Arg1 = register
 // Arg2 = size of element
-func ReadRegisterToOutputLocationExpression(register, size uint) LocationExpression {
+func ReadRegisterToOutputLocationExpression(register, size uint16) LocationExpression {
 	return LocationExpression{Opcode: OpReadUserRegisterToOutput, Arg1: register, Arg2: size}
 }
 
@@ -293,7 +293,7 @@ func ReadRegisterToOutputLocationExpression(register, size uint) LocationExpress
 // into a u64 which is then written to the output buffer
 // Arg1 = stack offset
 // Arg2 = size of element
-func ReadStackToOutputLocationExpression(offset, size uint) LocationExpression {
+func ReadStackToOutputLocationExpression(offset, size uint16) LocationExpression {
 	return LocationExpression{Opcode: OpReadUserStackToOutput, Arg1: offset, Arg2: size}
 }
 
@@ -303,7 +303,7 @@ func ReadStackToOutputLocationExpression(offset, size uint) LocationExpression {
 // back to the BPF parameter stack.
 // It should only be used for types of 8 bytes or less
 // Arg1 = size of value we're reading from the 8 byte address at the top of the stack
-func DereferenceLocationExpression(valueSize uint) LocationExpression {
+func DereferenceLocationExpression(valueSize uint16) LocationExpression {
 	if valueSize > 8 {
 		return LocationExpression{Opcode: OpDereferenceLarge, Arg1: valueSize, Arg2: (valueSize + 7) / 8}
 	}
@@ -316,7 +316,7 @@ func DereferenceLocationExpression(valueSize uint) LocationExpression {
 // directly to the output buffer.
 // It should only be used for types of 8 bytes or less
 // Arg1 = size of value we're reading from the 8 byte address at the top of the stack
-func DereferenceToOutputLocationExpression(valueSize uint) LocationExpression {
+func DereferenceToOutputLocationExpression(valueSize uint16) LocationExpression {
 	if valueSize > 8 {
 		return LocationExpression{Opcode: OpDereferenceLargeToOutput, Arg1: valueSize, Arg2: (valueSize + 7) / 8}
 	}
@@ -330,7 +330,7 @@ func DereferenceToOutputLocationExpression(valueSize uint) LocationExpression {
 // back to the BPF parameter stack.
 // Arg1 = size in bytes of value we're reading from the 8 byte address at the top of the stack
 // Arg2 = number of chunks (should be ({{.Arg1}} + 7) / 8)
-func DereferenceLargeLocationExpression(typeSize uint) LocationExpression {
+func DereferenceLargeLocationExpression(typeSize uint16) LocationExpression {
 	return LocationExpression{Opcode: OpDereferenceLarge, Arg1: typeSize, Arg2: (typeSize + 7) / 8}
 }
 
@@ -340,7 +340,7 @@ func DereferenceLargeLocationExpression(typeSize uint) LocationExpression {
 // This is safe to use for types larger than 8-bytes.
 // Arg1 = size in bytes of value we're reading from the 8 byte address at the top of the stack
 // Arg2 = number of chunks (should be ({{.Arg1}} + 7) / 8)
-func DereferenceLargeToOutputLocationExpression(typeSize uint) LocationExpression {
+func DereferenceLargeToOutputLocationExpression(typeSize uint16) LocationExpression {
 	return LocationExpression{Opcode: OpDereferenceLargeToOutput, Arg1: typeSize, Arg2: (typeSize + 7) / 8}
 }
 
@@ -350,7 +350,7 @@ func DereferenceLargeToOutputLocationExpression(typeSize uint) LocationExpressio
 // the output buffer.
 // Maximum limit (Arg1) should be set to the size of each element * max collection length
 // Arg1 = maximum limit on bytes read
-func DereferenceDynamicToOutputLocationExpression(readLimit uint) LocationExpression {
+func DereferenceDynamicToOutputLocationExpression(readLimit uint16) LocationExpression {
 	return LocationExpression{Opcode: OpDereferenceDynamicToOutput, Arg1: readLimit}
 }
 
@@ -361,13 +361,13 @@ func DereferenceDynamicToOutputLocationExpression(readLimit uint) LocationExpres
 // itself to be on the top of the stack.
 // Arg1 = string length limit
 func ReadStringToOutputLocationExpression(limit uint16) LocationExpression {
-	return LocationExpression{Opcode: OpReadStringToOutput, Arg1: uint(limit)}
+	return LocationExpression{Opcode: OpReadStringToOutput, Arg1: limit}
 }
 
 // ApplyOffsetLocationExpression creates an expression which
 // adds `offset` to the 8-byte address on the top of the bpf parameter stack.
 // Arg1 = uint value (offset) we're adding to the 8-byte address on top of the stack
-func ApplyOffsetLocationExpression(offset uint) LocationExpression {
+func ApplyOffsetLocationExpression(offset uint16) LocationExpression {
 	if offset == 0 {
 		return LocationExpression{Opcode: OpComment, Label: "apply_offset(0) == no-op"}
 	}
@@ -378,7 +378,7 @@ func ApplyOffsetLocationExpression(offset uint) LocationExpression {
 // writes to output `num_elements` elements, each of size `elementSize, from the top of the stack.
 // Arg1 = number of elements to pop
 // Arg2 = size of each element
-func PopLocationExpression(numElements, elementSize uint) LocationExpression {
+func PopLocationExpression(numElements, elementSize uint16) LocationExpression {
 	return LocationExpression{Opcode: OpPop, Arg1: numElements, Arg2: elementSize}
 }
 
@@ -391,7 +391,7 @@ func InsertLabel(label string) LocationExpression {
 // SetLimitEntry associates a collection identifier with the passed limit
 // Arg1 = limit to set
 // CollectionIdentifier = the collection that we're limiting
-func SetLimitEntry(collectionIdentifier string, limit uint) LocationExpression {
+func SetLimitEntry(collectionIdentifier string, limit uint16) LocationExpression {
 	return LocationExpression{Opcode: OpSetGlobalLimit, CollectionIdentifier: collectionIdentifier, Arg1: limit}
 }
 
@@ -400,7 +400,7 @@ func SetLimitEntry(collectionIdentifier string, limit uint) LocationExpression {
 // Arg1 = value to compare to global limit variable
 // CollectionIdentifier = the collection that we're limiting
 // Label = label to jump to if the value is equal to the global limit variable
-func JumpToLabelIfEqualToLimit(val uint, collectionIdentifier, label string) LocationExpression {
+func JumpToLabelIfEqualToLimit(val uint16, collectionIdentifier, label string) LocationExpression {
 	return LocationExpression{Opcode: OpJumpIfGreaterThanLimit, CollectionIdentifier: collectionIdentifier, Arg1: val, Label: label}
 }
 
@@ -425,7 +425,7 @@ func PrintStatement(format, arguments string) LocationExpression {
 func SetParameterIndexLocationExpression(index uint16) LocationExpression {
 	return LocationExpression{
 		Opcode: OpSetParameterIndex,
-		Arg1:   uint(index),
+		Arg1:   (index),
 	}
 }
 
@@ -445,9 +445,9 @@ func VerifierErrorLocationExpression() LocationExpression {
 // of capturing parameters from a running Go program
 type LocationExpression struct {
 	Opcode               LocationExpressionOpcode
-	Arg1                 uint
-	Arg2                 uint
-	Arg3                 uint
+	Arg1                 uint16
+	Arg2                 uint16
+	Arg3                 uint16
 	CollectionIdentifier string
 	Label                string
 	IncludedExpressions  []LocationExpression
